fix(scanner): Handle Sparse NCAs
There are some validation checks that try and open the filesystem to flag any errors that may exist. This is not possible with `sparse` NCAs so we skip or catch those errors correctly now. `Sparse` NCA files can still have their headers and hashes checked so overall validation is not effected.